+++ alembic/versions/2a1db4eb72e3_baseline.py
@@ -2,7 +2,7 @@
 
 Revision ID: 2a1db4eb72e3
 Revises: 
-Create Date: 2022-01-12 08:28:16.056281
+Create Date: 2022-01-14 23:40:10.120955
 
 """
 from alembic import op
@@ -33,8 +33,8 @@ def upgrade():
                     sa.Column('updated_at', sa.TIMESTAMP(),
                               server_default=sa.text('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'), nullable=False),
                     sa.PrimaryKeyConstraint('row_id'),
-                    sa.UniqueConstraint('name', 'symbol'),
-                    sa.UniqueConstraint('id')
+                    sa.UniqueConstraint('id'),
+                    sa.UniqueConstraint('name', 'symbol')
                     )
     op.create_table('token',
                     sa.Column('row_id', sa.BIGINT(), autoincrement=True, nullable=False),
@@ -52,14 +52,14 @@ def upgrade():
                               server_default=sa.text('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'), nullable=False),
                     sa.ForeignKeyConstraint(['blockchain_id'], ['blockchain.row_id'], ),
                     sa.PrimaryKeyConstraint('row_id'),
-                    sa.UniqueConstraint('blockchain_id', 'symbol'),
-                    sa.UniqueConstraint('id')
+                    sa.UniqueConstraint('id'),
+                    sa.UniqueConstraint('name', 'symbol', 'blockchain_id')
                     )
     op.create_table('conversion_fee',
                     sa.Column('row_id', sa.BIGINT(), autoincrement=True, nullable=False),
                     sa.Column('id', sa.VARCHAR(length=50), nullable=False),
-                    sa.Column('percentage_from_source', sa.DECIMAL(), nullable=True),
-                    sa.Column('amount', sa.DECIMAL(), nullable=True),
+                    sa.Column('percentage_from_source', sa.DECIMAL(precision=10, scale=5), nullable=False),
+                    sa.Column('amount', sa.DECIMAL(precision=50, scale=20), nullable=True),
                     sa.Column('token_id', sa.BIGINT(), nullable=True),
                     sa.Column('created_by', sa.VARCHAR(length=50), nullable=False),
                     sa.Column('created_at', sa.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP'),
@@ -77,8 +77,8 @@ def upgrade():
                     sa.Column('to_token_id', sa.BIGINT(), nullable=False),
                     sa.Column('conversion_fee_id', sa.BIGINT(), nullable=True),
                     sa.Column('is_enabled', sa.BOOLEAN(), nullable=True),
-                    sa.Column('min_value', sa.DECIMAL(), nullable=True),
-                    sa.Column('max_value', sa.DECIMAL(), nullable=True),
+                    sa.Column('min_value', sa.DECIMAL(precision=50, scale=20), nullable=True),
+                    sa.Column('max_value', sa.DECIMAL(precision=50, scale=20), nullable=True),
                     sa.Column('contract_address', sa.VARCHAR(length=250), nullable=False),
                     sa.Column('created_by', sa.VARCHAR(length=50), nullable=False),
                     sa.Column('created_at', sa.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP'),
@@ -116,9 +116,9 @@ def upgrade():
                     sa.Column('row_id', sa.BIGINT(), autoincrement=True, nullable=False),
                     sa.Column('id', sa.VARCHAR(length=50), nullable=False),
                     sa.Column('wallet_pair_id', sa.BIGINT(), nullable=False),
-                    sa.Column('deposit_amount', sa.DECIMAL(), nullable=False),
-                    sa.Column('claim_amount', sa.DECIMAL(), nullable=True),
-                    sa.Column('fee_amount', sa.DECIMAL(), nullable=True),
+                    sa.Column('deposit_amount', sa.DECIMAL(precision=50, scale=20), nullable=False),
+                    sa.Column('claim_amount', sa.DECIMAL(precision=50, scale=20), nullable=True),
+                    sa.Column('fee_amount', sa.DECIMAL(precision=50, scale=20), nullable=True),
                     sa.Column('status', sa.VARCHAR(length=30), nullable=False),
                     sa.Column('claim_signature', sa.VARCHAR(length=250), nullable=True),
                     sa.Column('created_by', sa.VARCHAR(length=50), nullable=False),
@@ -153,7 +153,7 @@ def upgrade():
                     sa.Column('transaction_visibility', sa.VARCHAR(length=30), nullable=True),
                     sa.Column('transaction_operation', sa.VARCHAR(length=30), nullable=True),
                     sa.Column('transaction_hash', sa.VARCHAR(length=250), nullable=True),
-                    sa.Column('transaction_amount', sa.DECIMAL(), nullable=True),
+                    sa.Column('transaction_amount', sa.DECIMAL(precision=50, scale=20), nullable=True),
                     sa.Column('status', sa.VARCHAR(length=30), nullable=True),
                     sa.Column('created_by', sa.VARCHAR(length=50), nullable=False),
                     sa.Column('created_at', sa.TIMESTAMP(), server_default=sa.text('CURRENT_TIMESTAMP'),