+++ domain/entities/blockchain.py
@@ -1,10 +1,13 @@
+from datetime import date
+
 from constants.entity import BlockchainEntities
+from utils.general import datetime_to_str
 
 
 class Blockchain:
     def __init__(self, id: str, name: str, description: str, symbol: str, logo: str, chain_id: str,
-                 block_confirmation: int, is_extension_available: bool, created_by: str, created_at: str,
-                 updated_at: str):
+                 block_confirmation: int, is_extension_available: bool, created_by: str, created_at: date,
+                 updated_at: date):
         self.id = id
         self.name = name
         self.description = description
@@ -14,8 +17,8 @@ class Blockchain:
         self.block_confirmation = block_confirmation
         self.is_extension_available = is_extension_available
         self.created_by = created_by
-        self.created_at = created_at
-        self.updated_at = updated_at
+        self.created_at = datetime_to_str(created_at)
+        self.updated_at = datetime_to_str(updated_at)
 
     def to_dict(self):
         return {