+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Released]
+## 2022-01-17 12:00:00 
+
+### Added
+- Added the token api, testcase and swagger documentation
+
 ## 2022-01-12 16:00:00 
 
 ### Added
 - Setup the base code structures and utils
 - Added the blockchain api and testcase
 - Added the swagger documentation
-- 
+